requirements.txt
pytelegrambotapi
speedtest-cli
requests
playsound
psutil
opencv-python
pillow